car2data_project/requirements/base.txt
# Framework principal
Django==4.2.7
djangorestframework==3.14.0

# Base de datos - PyMySQL es más compatible con Windows
PyMySQL==1.1.0
cryptography==41.0.7

# Procesamiento de archivos
PyPDF2==3.0.1
python-docx==1.1.0
Pillow>=10.0.0  # Versión flexible para mejor compatibilidad

# IA y procesamiento
openai==1.3.3
google-generativeai==0.3.0

# Validación y formularios
django-crispy-forms==2.0
crispy-bootstrap5==0.7

# Seguridad y autenticación
django-cors-headers==4.3.1
django-extensions==3.2.3
python-decouple==3.8

# Utilidades
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
boto3==1.29.7

# Fechas y tiempo
python-dateutil==2.8.2

# Procesamiento de texto
python-magic-bin==0.4.14  # Para Windows